Fix CocoaPods Build
Part C.
- Ensure schemes contain the correct build entries.
- Make sure that CocoaPods build customizations affect all relevant build items (there were some (at least the Pods build root) left over which excluded ARM64; PodFile edited to correct).
- Add CocoaPods build customizations to the overall build root (Cicerone; the Pod build customization via hook in the PodFile takes care of the Pods build root. See item 2).
Note that there is no real need to force NATIVE_ARCH to be the sole value of "Valid Architectures", but was done as a troubleshoot fix/catch-all to make sure the build would work with a "valid" selected architecture. The intent is that this will be reverted later, to allow CI builds to work.
Also note that I mention ARM64 here and in the last commit, however it is unclear if that was ever the issue; the contribution of what seemed to be this ARM64 exclusion element to the build issues, could have been errant and coincidental.